UI_Test_01/seleweb/requirements.txt
PyYAML == 5.1.2
selenium == 3.14.1
xlrd == 1.1.0
